\name{gaggleInit}
\alias{gaggleInit}
\title{Initialize the R Goose, registering with the Gaggle?}
\description{You must call this function in order to register the R
  Goose with the already-running Gaggle Boss.  This makes all subsequent
  communication possible.
}
\usage{gaggleInit ()}
\value{Returns NULL.}

\seealso{
  \code{\link{broadcast}}.
  \code{\link{getNameList}}.
  \code{\link{getMatrix}}.
  \code{\link{getNetwork}}.
  \code{\link{getCluster}}.
  \code{\link{getHashMap}}.
  }

\keyword{manip}

