\name{getNameList}
\alias{getNameList}
\title{Create an R variable from Most Recently Broadcast NameList}
\description{After a nameList is broadcast to the R goose, one additional step is required of the R
             user before it is available as a character list variable in R.  (This step may be eliminated in
             the future.)}
\usage{names <- getNameList ()}
\arguments{
  }

\value{Returns a list of names.}

\seealso{
  \code{\link{getMatrix}}.
  \code{\link{getNetwork}}.
  }
\keyword{manip}

