\name{getTuple}
\alias{getTuple}
\title{Create an R variable from Most Recently Broadcast Tuple}
\description{After a tuple is broadcast to the R goose, this method makes it available as an R environment object.}
\usage{env <- getTuple ()}
\arguments{
  }

\value{Returns an environment with a title, and then the usual keys and values.}

\seealso{
  \code{\link{newTuple}}.
  \code{\link{getNameList}}.
  \code{\link{getCluster}}.
  \code{\link{getMatrix}}.
  \code{\link{getNetwork}}.
  }
\keyword{manip}
\keyword{tuple}
\keyword{gaggle}
\keyword{broadcast}

